+++ ext/SpaceDataModelExt.jl
@@ -1,14 +1,4 @@
-using SpaceDataModel
-using SpaceDataModel: Product, AbstractDataVariable, AbstractProduct, AbstractDataSet
+using SpaceDataModel: AbstractDataVariable
 
-mappable(x::Product) = (x,)
 plottype(x::AbstractDataVariable) = isspectrogram(x) ? SpecPlot : LinesPlot
-plottype(::AbstractProduct) = FunctionPlot
-plottype(::AbstractDataSet) = MultiPlot
-
 makie_x(da::AbstractDataVariable) = makie_t2x(times(da))
-
-# Makie.convert_arguments(::Type{<:LinesPlot}, da::AbstractDataVariable; kwargs...) = plot2spec(LinesPlot, da; kwargs...)
-
-transform(x::AbstractProduct, args...) = DimArray ∘ x
-transform(x::AbstractDataSet, args...) = x